bit refactoring
