+++ packages/connectivity/src/scp-cf/destination/destination-from-service.ts
@@ -1,39 +1,24 @@
-import { createLogger } from '@sap-cloud-sdk/util';
-import { JwtPayload } from '../jsonwebtoken-type';
-import { decodeJwt, isUserToken, JwtPair, verifyJwt } from '../jwt';
-import { jwtBearerToken, serviceToken } from '../token-accessor';
-import { addProxyConfigurationOnPrem } from '../connectivity-service';
-import {
-  getDestinationService,
-  getDestinationServiceCredentialsList
-} from '../environment-accessor';
-import { isIdenticalTenant } from '../tenant';
-import { DestinationServiceCredentials } from '../environment-accessor-types';
-import { exchangeToken, isTokenExchangeEnabled } from '../identity-service';
-import { getSubdomainAndZoneId } from '../xsuaa-service';
-import { userId } from '../user';
-import { Destination } from './destination-service-types';
-import {
-  alwaysProvider,
-  alwaysSubscriber,
-  subscriberFirst
-} from './destination-selection-strategies';
-import {
-  DestinationFetchOptions,
-  DestinationsByType
-} from './destination-accessor-types';
+import {createLogger} from '@sap-cloud-sdk/util';
+import {JwtPayload} from '../jsonwebtoken-type';
+import {decodeJwt, isUserToken, JwtPair, verifyJwt} from '../jwt';
+import {jwtBearerToken, serviceToken} from '../token-accessor';
+import {addProxyConfigurationOnPrem} from '../connectivity-service';
+import {getDestinationService, getDestinationServiceCredentialsList} from '../environment-accessor';
+import {isIdenticalTenant} from '../tenant';
+import {DestinationServiceCredentials} from '../environment-accessor-types';
+import {exchangeToken, isTokenExchangeEnabled} from '../identity-service';
+import {getSubdomainAndZoneId} from '../xsuaa-service';
+import {Destination} from './destination-service-types';
+import {alwaysProvider, alwaysSubscriber, subscriberFirst} from './destination-selection-strategies';
+import {DestinationFetchOptions, DestinationsByType} from './destination-accessor-types';
 import {
   AuthAndExchangeTokens,
   fetchDestination,
   fetchInstanceDestinations,
   fetchSubaccountDestinations
 } from './destination-service';
-import { destinationCache, IsolationStrategy } from './destination-cache';
-import {
-  addProxyConfigurationInternet,
-  ProxyStrategy,
-  proxyStrategy
-} from './proxy-util';
+import {destinationCache, getIsolationStrategy} from './destination-cache';
+import {addProxyConfigurationInternet, ProxyStrategy, proxyStrategy} from './proxy-util';
 
 type DestinationOrigin = 'subscriber' | 'provider';
 
@@ -579,10 +564,3 @@ Possible alternatives for such technical user authentication are BasicAuthentica
   }
 }
 
-function getIsolationStrategy(jwt: JwtPayload | undefined): IsolationStrategy {
-  if (jwt && userId(jwt)) {
-    return IsolationStrategy.Tenant_User;
-  }
-
-  return IsolationStrategy.Tenant;
-}
